File Upload, ASP Tokens, View Updates, Blog Management Added, Product Report Added

+++ DiscountManagement.Domain/CustomerDiscountAgg/CustomerDiscount.cs
@@ -9,9 +9,9 @@ namespace DiscountManagement.Domain.CustomerDiscountAgg
         public DateTime StartDate { get; private set; }
         public DateTime EndDate { get; private set; }
         public string Reason { get; private set; }
-        public decimal DiscountPercentage { get; private set; }
+        public int DiscountPercentage { get; private set; }
 
-        public CustomerDiscount(long productId, DateTime startDate, DateTime endDate, string reason, decimal discountPercentage)
+        public CustomerDiscount(long productId, DateTime startDate, DateTime endDate, string reason, int discountPercentage)
         {
             ProductId = productId;
             StartDate = startDate;
@@ -19,7 +19,7 @@ namespace DiscountManagement.Domain.CustomerDiscountAgg
             Reason = reason;
             DiscountPercentage = discountPercentage;
         }
-        public void Edit(long productId, DateTime startDate, DateTime endDate, string reason, decimal discountPercentage)
+        public void Edit(long productId, DateTime startDate, DateTime endDate, string reason, int discountPercentage)
         {
             ProductId = productId;
             StartDate = startDate;
